Revert "Move UI components into UI"
This reverts commit 3379e61246cde635c296d31322b71e63395a5cd4.
