+++ js/src/modals/SMSVerification/store.js
@@ -21,8 +21,9 @@ import { sha3 } from '~/api/util/sha3';
 import Contracts from '~/contracts';
 
 import { checkIfVerified, checkIfRequested, awaitPuzzle } from '~/contracts/sms-verification';
-import { postToServer } from '~/3rdparty/sms-verification';
-import { checkIfTxFailed, waitForConfirmations } from '~/util/tx';
+import { postToServer } from '../../3rdparty/sms-verification';
+import checkIfTxFailed from '../../util/check-if-tx-failed';
+import waitForConfirmations from '../../util/wait-for-block-confirmations';
 
 export const LOADING = 'fetching-contract';
 export const QUERY_DATA = 'query-data';
